feat: fold refresh into atomWithCache via { type: 'refresh' } action
