+++ src/index.ts
@@ -1,2 +1,1 @@
 export { atomWithCache } from './atomWithCache.js';
-export { atomWithCacheAndRefresh } from './atomWithCacheAndRefresh.js';
